Update patch set 2
Patch Set 2:

Good idea!
In the beginning I was a bit confused that I have to look for options at two places. So it would be better to also put "Show untracked files" and "Configure Preferences" in the toolbar. It would also be nice if there would be preferences for adding signed of and gerrit change id automatically.

Patch-set: 2