Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(3 inline comments)

I like the idea ...

I think we should go one step further:
- also move "Configure preferences" to the toolbar
- make "Show untracked files" a toggle button in the toolbar

I feel the icon for sign-off should contain something like a check mark. The icon for amending is hard to understand, the blueish blur is disturbing.

Maybe we could in addition add a menu in the style of a view menu (don't know if that fits to Eclipse UI standards). For beginners a menu is easier to understand. And the little triangle for opening the menu wouldn't clutter the UI so it wouldn't disturb experienced users.

Patch-set: 2