Update patch set 1
Patch Set 1:

After using the new History view color palette for the last month, I've found the first blue color easier to look at then the first green color.  It is important that the first color is the palette is appealing since it will be used exclusively when looking at history that is linear with no branches.

Patch-set: 1