Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

I don't see value in basing this wizard on project-level, I would prefer repository-level as this 
- shortens the list
- requires less clicks
- fits better to git semantics (commits intrinsically may contain modifications across complete repo)
- project based wizard ignores non-workspace level files which are outside projects, this is no problem with repo level
- still result can show projects

We could also remove the checkboxes and just check if destination has been set to decide which repos should be sychronized.

The wizard should remember the configuration from last run, this would be handy if I run synchronize, then do some changes and want to rerun it with the same configuration to check the effect of my changes.

Patch-set: 4