Fix NumberFormatException on "Create Branch"
When a repository contains a branch or tag which includes many numbers
(e.g. stable-1-0-0_2011-05-25), it was impossible to create a new branch
because of a NumberFormatException:

  java.lang.NumberFormatException: For input string: "10020110525"

The cause was in CommonUtils.STRING_ASCENDING_COMPARATOR used from here:

  org.eclipse.egit.ui.internal.repository.CreateBranchPage.createControl(CreateBranchPage.java:182)

The (not optimal) fix is to use BigInteger instead of Integer.parseInt.

Change-Id: Idf7f812cc1728bcb8098652bf60c1bb7141ee537
Signed-off-by: Robin Stocker <robin@nibor.org>