Fix CommonUtils.STRING_ASCENDING_COMPARATOR issues
The comparator provides the means to sort a list of strings using a
natural sort order (i.e. "2" before "10"). It had two problems.

The first one was that it could not handle strings that contained many
numbers (when all the numbers concatenated exceeded Integer.MAX_VALUE).
This lead to a NumberFormatException. Because the comparator is used in
CreateBranchPage, it was impossible to create a branch in a repository
that had a ref with such a name (e.g. stable-1-0-0_2011-05-25).

The second one was that it resulted in the wrong order for some inputs.
For example, the following inputs would not be sorted as listed here:

  project-1-0-0-final
  project-1-0-1-beta

With the implementation of this change, it behaves correctly for these
two problems.

Change-Id: Ie5edf268889f43c2a4497669aa1700f790cd2cdd
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>