Update patch set 1
Patch Set 1:

The class extends FormPage which extends EditorPart and so I think CommitEditorPage is an appropriate name.  Although it is read-only now, the plan is to make parts of it editable in the future (such as adding/editing notes https://bugs.eclipse.org/bugs/show_bug.cgi?id=347034).

The reason I felt comfortable committing this change myself in a short time-frame was that is was a UI-quirk bug in an unreleased feature that I had developed and so I considered it trivial which the EGit contributor guide seems to support: "let non-trivial changes be in review for at least 24 hours"

Patch-set: 1