[RFC] Fix GitRepositoriesViewFetchAndPushTest
ContextMenuHelper was removed from GitRepositoriesViewFetchAndPushTest
because it was assumed that ContextMenuHelper causes the test failures.
Instead of ContextMenuHelper the context menu is triggered by using
the SWTBot contextMenu method. To make this work the setEnabled method
of FetchConfiguredRemoteCommand was changed to use
ISources.ACTIVE_CURRENT_SELECTION_NAME instead of
ISources.ACTIVE_MENU_SELECTION_NAME. Otherwise the test fails because
FetchConfiguredRemoteCommand is not enanbled when clicking the Fetch
context menu entry.

Change-Id: I49e1944b780225ae733a05b84e1e263b51d82f84
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>