Generate pre-built help index - local PDE Build / PDE Export support.
This commit adds a pre-built help index when building with PDE build
(eg in the local workspace). As a follow-up of this change, the help
index building needs to be integrated for Tycho, or the added setting
in plugin.xml will lead to an error message when searching help the
first time.

Bug: 350861
Change-Id: Ie438e5a642eaf1e50c9482898822bc5376cbe377