bug 347073 - clarify difference btw. remove and delete repository
The context menu of the repository view has the menu items "Remove
Repository" and "Delete Repository". "Remove Repository" removes a
repository from the repository view whereas "Delete Repository" deletes
a local [clone of a] repository from disk.

The difference btw. these two is not obvious enough. It was suggested to
adda tooltip to the context menu items but this appears not to be
technically possible. 

The patch proposes to rename the context menu item "Remove Repository"
to "Remove Repository from View".

Change-Id: Ic9ccd5c9f827a84c6a434c6a07bf5fbdeede7f46
Signed-off-by: Adrian Goerler <adrian.goerler@sap.com>