Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This patch is reacting to the state of the SWT item, but it is a handler for a command.  If you ever show this command in 2 places (a toolbar and a menu, for example) the state will get out of sync.

See http://blog.eclipse-tips.com/2009/03/commands-part-6-toggle-radio-menu.html for an example of implementing checked or radio state using commands.

Patch-set: 1