Save stagingview commit message to look for updates
By saving the initial commit message we can check if the user
has updated the text or not so we can tell when it should be
recalculated.  If the user changes the commit message then
keep it otherwise re-calculate it.

This should allow commit message contributors like the mylyn
plugin to supply the current commit message however since the
commit message is only updated on refresh and not task activation
a change to the current task will only be reflected on refresh.

Bug: 348744
Change-Id: I4f7a12b5172a97a68918f33490096e0131a9af66