Fix wrong destination ref proposal in PushWizard
When the user entered "tags/myTag" into the source field the PushWizard
proposed "tags/myTag" in the destination field. This is then expanded to
"refs/tags/tags/myTag" by JGit. The analogous behavior was there if the
user entered heads/myBranch.

This commit fixes the bug; now the only the name after "tags/" resp.
"heads/" is proposed in the destination field.


Bug: 341069
Change-Id: I108ff09ef39c3aefa5676c7724af2e42e3600172
Signed-off-by: Stefan Lay <stefan.lay@sap.com>