Fix remote tracking branch update on push
Pushing via "Team->Push to upstream" and pushing via executing "Push" on
a remote's push node now updates the related remote tracking branch.
A third constructor was added to PushOperationUI to enable pushing on a
given remote name. The constructor of PushOperation was changed to take
the remote name instead of a remote config.
The new constructor in PushOperationUI was required because JGit's
PushCommand only updates the remote tracking branch if a remote name is
given. No update is performed if a push uri is provided.

Depends on JGit change
I2bc648abc517b1d01b2de15d383423ace2081e72

Bug: 317411
Change-Id: I42b0ac94be979b36cb3f6d78141f6c65a5e5efbf
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>