Update patch set 1
Patch Set 1:

The thing is that the view doesn't "clear" itself if you select a resource that's not under version control. This is not necessarily a "bad" idea as the view would be forced to reindex very often if you kept clicking between a resource that was under version control and one that wasn't.

On the other hand, this means that the view would continue to show information about the last valid repository even when all the projects versioned under that repository were closed. So now the view is showing information about a repository that's no longer applicable for the current workspace.

Patch-set: 1