[RFC] Use IStorageMerger to perform certain merges
** DO NOT SUBMIT **

In order to support other file formats that need to be merged in a
special way, we should introduce a new content merger that uses the team
framework (especially IStorageMerger and IStreamMerger) to merge the
contents of files.
For the first pass, only CherryPick is used as example.
* Support for the other merge cases should be part of this change.
* Test case that provides a mock IStorageMerger


Depends on I7817e212
Change-Id: I2b3ec2771cc1e901f3615b8c2909e4f984dcae5d
Signed-off-by: Benjamin Muskalla <benjamin.muskalla@tasktop.com>