Do not downcast provided evaluation context as an EvaluationContext
The context that's passed into handlers are not guaranteed to be an
EvaluationContext so performing an instanceof check for that type
may potentially go nowhere. The fix is to use HandlerUtil's
getVariable(Object, String) method for retrieving variables from the
context. This lets the framework handle the processing of the
context instead of having EGit make arbitrary assumptions about the
type of the provided context.

Bug: 352382
Change-Id: I835f5629ba6e3b7853b2233e1ddb95453e01332f
Signed-off-by: Remy Suen <remysuen@ca.ibm.com>