MultiPullResultDialog: Fix overall status column
It would always say "OK" even when the rebase or merge failed. It now
displays "Failed" when the status is not successful.

Also, the logic for rebase was inverted, leading to the image indicating
failure when the rebase was successful.

Depends on JGit change Icfc9a9c76763f8a777087a1262c8d6ad251a9068 for
isSuccessful.

Change-Id: Id3bb809c83464c20b20a93aeae048f9a2edbea63
Signed-off-by: Robin Stocker <robin@nibor.org>