Update patch set 5
Patch Set 5:

It seems that the issue occurs when after the dnd-operation a folder or the whole tree should not be shown anymore.

Could you please check how the "Next Difference" action handles the case when a folder or the whole tree is not available anymore after a dnd-operation?

Patch-set: 5