Don't use saved dialog settings when pull results are empty.
The PullResultDialog does not showing trees or tables
when there is nothing to pull so the dialog will typically
open too large for the no updates message when it has
been previously resized when actual results were shown.

The PullResultDialog will open in a smaller packed mode
only in the case where there are no fetch, merge, or rebase
results.

Change-Id: I1895c378a71ef4fe5bef81424c633c9d8de9b309
Signed-off-by: Kevin Sawicki <kevin@github.com>