Show branch status in label decoration of project
The branch status is shown as a "*" in front of the branch name if the
current branch is different from the remote tracking branch. This is
useful as a visual indication that the branch needs to be pushed (or
merged/rebased).

For example, if there are unpushed commits on master, the project is now
shown like this (similar to a modified file that needs to be saved):

  project [repo *master]

In C Git, information about the tracking branch is displayed in "git
status" output. EGit currently doesn't have any indication that a branch
needs to be pushed, apart from the labels in History View. When there
are multiple repositories involved, it can easily happen that a push is
forgotten. This change helps prevent that.

Change-Id: I94a78ef169b4ae85f7061f7a99f8bed5e8a8838b
Signed-off-by: Robin Stocker <robin@nibor.org>