Update patch set 3
Patch Set 3:

Guys, I need some help here with the clone destination directory.

The problem is this: GitProjectSetCapability is in core (not ui), because it is configured in GitProviderType, which in turn is configured in the "org.eclipse.team.core.repository" extension point in core/plugin.xml.

So showing a dialog is not possible from GitProjectSetCapability, right?

Ok, so maybe we can just use the "default_repository_dir" preference? The problem with this is that this preference is also in ui. Can we move it to core? I don't know, it's used in GitPreferenceRoot via the preference store.

Patch-set: 3