Update patch set 3
Patch Set 3:

nice enhancement

I think the default should be to show reflog for HEAD (equivalent to what "git reflog" displays which defaults to "git reflog show HEAD") this will also display branch switches.

Another nice addition would be to add a toggle "Link with editor and selection" so that the view always shows the reflog for the currently checked-out branch of the repo the selected resource is in.

Somewhen soon we should add some tests for this view.

Patch-set: 3