Use resource encoding in getFileCachedRevisionTypedElement
This fixes problems in compare actions that use index as base (Compare
index with HEAD, Staging View and Synchronize View).

When the workspace encoding is different from the encoding of the
resource, the index side of the diff would be decoded incorrectly (while
the other side would be decoded correctly), leading to spurious diffs.

Bug: 345160
Change-Id: If84334131ad53732aeac595e43ac0972de359628
Signed-off-by: Robin Stocker <robin@nibor.org>