Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Why should we not want to allow for a merge configuration using a local branch as base?

Currently we do the same as

git branch newbranch master --track

(where "master" is a local branch) if the user chooses one of the option "merge" or "rebase",

and the same as

git branch newbranch master

if the user chooses "none".

Patch-set: 1