Update patch set 4
Patch Set 4:

tried with your sample model:

we should not show the dialog "Selection Adjustment Required" if replace wouldn't touch the involved files as there is no diff. E.g. if the selected resource including its dependent model files are unchanged and the user clicks "Replace With > HEAD Revision" there is no need to show this dialog as the operation is a null op anyway.

Patch-set: 4