Diff node should notify listeners when the input changes
The text merge viewers need to be notified when changes to a
compare input are saved. The text viewer can be notified of
changes from the diff node. Hence, when the compare input itself
has been notified that its been changed, this change event must
be forwarded to the diff node so that the text viewer can be
updated and reflect the differences of the newly changed compare
input. 

Bug: 352576
Change-Id: I2f83544f271fea4a4e21f346218a75ef67bae623