Implement project rename and move for some cases
We can now move and rename projects within a repository for a number
of cases. Moving a project to a subdirectory does not work and since
Eclipse actually destroys the project, we simply forbid it for now
(Platform bug 307140). The bug is only worked around in Eclipse 3.7,
i.e. forbidden in the UI.

This also adds tests for move and delete of folders and files. We are
quite inconsistent with respect to how we handle the index so the
current behavior may not be the final behavior.

We cannot move projects with repositories inside them in this
version and so we forbid it. Disconnect and share manually
to accomplish this.

Bug: 334112
Change-Id: I634d5b44bc7fc6a7453b5d718733c9f8f3542840
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>