Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

- error messages improved
- don't clone if destination directory exists (with detailed error message)

It still requires a compatibility layer. Unfortunately, providing it as an optional dependency not a good solution from the user's PoV: if a compatibility layer is missing there will be no feedback provided by team framework. Should be moved to a feature patch or separate feature to not mandating a compatibility layer for EGit.

Patch-set: 4