Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

The absolute dependency on the runtime bundle should not be pushed through. We should, if we're working around the bug, make it an optional requirement so that if present, it's available, but if not, we don't need it. IIRC the only thing that will fail is the export of the PSF if this dependency isn't satisfied, and that can be fixed upstream.

Patch-set: 2