Update patch set 2
Patch Set 2: I would prefer that you didn't submit this; No score

and, now Shawn removed the eclipse project files so the provide psf isn't usable even if the URLs get fixed.

Could we get the import to run mvn eclipse:eclipse in a clean fashion? That would be useful in other settings than Git. We could drop that psf file from the patch and it would almost be useful if it weren't because we hsve not
decent way of creating project set files. This patch enables PSF export, but it doesn't work (no project serializer).

Patch-set: 2