[search] Prevent overly large repository area in dialog.
The search dialog table that shows the Git repositories was
grabbing an excess amount of space causing the dialog to open
larger than the screen height with no scrollbars on the table.
This was apparent if you had ~50 repositories configured in the
Git Repositories view.

This was due to the fact that the SearchDialog class sets the
layout data on the main control of the search page after it is
created which overrides the locally set data.

The fix prevents setting the layout data on the main control
after is had already been set by using a simple subclass of
Composite.

Change-Id: I3d60524054e92ef48f0102cb4d0c855d57162a86
Signed-off-by: Kevin Sawicki <kevin@github.com>