Update patch set 2
Patch Set 2:

You could use the same message as we have in CommitAction_MergeHeadErrorMessage

But I really wonder if there is a way to help the user to resolve this state. Looking at bug 352243, it seems the user has no clue how he got into this state and what to do.

Patch-set: 2