Improve handling of notes commits in History View
When he FETCH_HEAD pointed to refs/notes/* the notes commits were shown
in History View if the "Show all Branches and Tags" button was selected.
Normally this is not what the user expects.

With this commit the notes commits are not shown anymore per default.
Instead there is now a new action and a corresponding preference which
controls if the notes commits are shown


Change-Id: I51b746c8ba8f7932156bd00362b9e6bcb0d755cf
Signed-off-by: Stefan Lay <stefan.lay@sap.com>