[DRAFT] task activation integration support
This change enables the automatic creation or checkout of branches
according to the activated mylyn task.
If a new task is activated, the branch to create can be chosen with
a dialog. On task deactivation, a Project Set file is stored besides
the Mylyn context zip file. This file is read on task activation
to determine the branches and repositories.

Change-Id: Ife371e3b33ea477ed8fec24c83548d25c1fef9df
Signed-off-by: Manuel Doninger <manuel.doninger@googlemail.com>