Ignore undefined variables when retrieving values from the context
When variables are retrieved from the evaluation context, it can
be null or it may be an undefined value. When a value is
undefined, we want our selection searching algorithm to cascade
down and request a value from a different variable. This ensures
that both null and undefined values are ignored and that our
handlers can ask other a different variable for a value instead
of taking the undefined value as a valid return value.

Bug: 356647
Change-Id: Ib349fdd69288c46173101b9d4ae3d6fbfb9d69c0