Clone Wizard: fix problem with review notes
When cloning a git repository hosted in gerrit (check option
"Configure Push to Gerrit Code Review" in Clone Wizard) problems occur
if the git repository does not contain review notes.
Reason for the failing fetch is the addition of refs/notes/review to the
fetch spec. This causes the fetch to fail if review notes are not
available.
The problem is fixed by adding refs/notes/* to the fetch spec instead.
This does not cause the fetch command to abort if no review notes are
available.

Bug: 357781
Change-Id: Ica1349441626c9aafbdd580bc6ed8f3e65271d64
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>