Enable 'Pull' operation when selecting multiple repositories
The 'Pull' operation is already enabled when selecting multiple
projects spread across different repositories due to the work
of commit 4197370f. To enable this feature for the 'Git
Repositories' view, similar work must be done in the handler to
analyze that every single element in the selection corresponds
to a repository pointing at valid head reference. If this
criteria is satisfied, the 'Pull' operation should be enabled.

Bug: 356781
Change-Id: I0bd3779f885a6df77239da8c74e214e5b53d8768