Implement project rename and move for some cases
We can now move and rename projects within a repository for a number
of cases. Moving a project to a subdirectory does not work and since
Eclipse actually destroys the project, we simply forbid it for now
(Platform bug 307140). The bug is only worked around in Eclipse 3.7,
i.e. forbidden in the UI.

This also adds tests for move and delete of folders and files. We are
quite inconsistent with respect to how we handle the index, i.e.
sometime we update the index to reflect the change and sometimes we do
not, so the current behavior may not be the final behavior.

RFC: Moving repositories require disconnect/connect and this is, I think
not reliable yet.

This version of the patch does not try to evade bug 307140

Requires JGit patch: I7425da91c0752ae82484e3c29d21b57402d30c61

Bug: 334112
Change-Id: I634d5b44bc7fc6a7453b5d718733c9f8f3542840
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>