Decorate lightweight tags with checked out overlay.
Lightweight tags that point to the current detached HEAD commit
will now display the checked out overlay in the top right corner
similar to branches and annotated tags.

This also removes the 'L' overlay currently shown for lightweight
tags since it was displayed in the top right corner and would be
invisible when the checked out overlay was also drawn.

The 'L' looked abnormal when in tested in other corners so instead
a new icon was added for annotated tags and lightweight tags will
use the simple tag icon.

The new annotated tag icon is the standard tag icon with an
author overlay on the right side denoting that the tag has a
known tagger and message.

Bug: 357756
Change-Id: Ib505c58c42347ffaa612bada1f3259b1e51d0346
Signed-off-by: Kevin Sawicki <kevin@github.com>