Introduce FilterPath class for following renames
Introduced a FilterPath class that holds a path
and the information whether the path is a
regular file or a directory.

This will be useful to implement the 'Follow Renames'
feature that needs to be effective only if filtered
path are actually one single regular file.

Had to change the author tag from
Benjamin <benje@linkeo.com> to Carsten Pfeiffer, because I couldn't push
to Gerrit otherwise.

Change-Id: I2270ec8b68886b3715f2e07099d96486e5dcb504