Added the 'Follow Renames' feature to the history page.
This adds a dedicated BooleanPrefAction, which is added to the 'Show'
sub-menu. When activated, an history page on one single regular file
sets a proper FollowFilter instance as TreeFilter for the current
walk.

The fileViewer now correctly displays those files that are related to
the currently selected commit and file, taking renaming of files into
account).

AlsoBy: Benjamin Gandon <benje@linkeo.com>
Change-Id: Id810e22db3f415a8471a7c4c80d9abe9fcaeb21f