Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

(3 inline comments)

When a commit is selected which moved/renamed a file the file list in the "Revision Details" pane of the History View could show the file as one "rename" entry (instead of a deleted file and a newly added file) and give both the old and the new path (similar to what "git log --stat --follow path/to/renamed/file" shows for renamed files), then also a new decorator icon should be used for the renamed entry. Of course this may deserve a separate change.

Patch-set: 4