Restrict synchronization cache updates to specific resources
If a synchronization operation is only asking a specific set of
resources to be updated, only that particular set of resources
should be filtered out for an update of the synchronization
cache. This removes the need to walk the entire tree when the
user has only made changes to a small number of resources.

Change-Id: I36267891fac1e33ba38b98dd121f5877c5da54ee
Signed-off-by: Remy Suen <remysuen@ca.ibm.com>