Fix exception when no review notes are available
When there has not yet been done a code review on the remote Gerrit
fetch throws an exception when there is the refspec
refs/notes/review:refs/notes/review. Since it is standard git behavior
to expect a ref on the remote if there is no wildcard in the refspec we
do not add this refspec in this case. 

Bug: 357781
Change-Id: I14a65273a43ae53640d822b26f44e8afc558a4bd
Signed-off-by: Stefan Lay <stefan.lay@sap.com>