Handle excluded resources properly in 'Git Staging' view
As the 'Git Staging' view is intended to be a mirror of the 'git
status' command in the command line, it should use the standard
FileTreeIterator from JGit for walking over the working
directory and index. Using EGit APIs that interact with
Eclipse's resources layer will not work correctly because
IContainer's members(*) method will not return excluded
resources. As the command line does not consider excluded
resources any differently from regular files, the 'Git Staging'
view should follow the same guideline and not deviate in its
behaviour and display of information.

Bug: 362319
Change-Id: Id875f858a5fa3feff3846ea368ae07e76a004d11
Signed-off-by: Remy Suen <remysuen@ca.ibm.com>