Update patch set 3
Patch Set 3:

You're right, Dariusz, thanks for catching this. How embarrassing. :|

What is interesting is that it _does_ work if I try org.eclipse.jgit.console and org.eclipse.jgit.iplog. It seems that the path filtering optimization causes the org.eclipse.jgit folder to not be traversed by the walk for some reason.

Patch-set: 3