Update patch set 3
Patch Set 3: (2 inline comments)

Dariusz, if you plan to fix the breakpoint problem in this change set, please document this in the commit message.

With the old code, making changes to a file and saving it will cause it to be shown in the 'Synchronize' view. This would stop working very quickly afterwards (in the sense that it would no longer refresh properly). With the new code, this feature won't even work once.

Patch-set: 3