Do not create a filter for empty string paths
When the merge compare editor input is created, it uses the
resources selected by the user to determine what the tree
structure of compared resources should be. If a selected project
is both the parent of the .git folder as well as the selected
root resource to perform the merge for, an empty string will be
determined as the filter to use for generating the differences.
In this case, a path filter should not be created for the empty
string as it is unnecessary as this would be identical to simply
iterating over all of the files in the repository. 

Bug: 347551
Change-Id: I92ea4d4fa95edd8d9b2fef7a820cc0adacb520c7
Signed-off-by: Remy Suen <remysuen@ca.ibm.com>