Update patch set 2
Patch Set 2: No score

Dariusz, if you find that the resource optimization is the problem between renamed files and resource changes outside of Eclipse not showing up in the view, please remove the optimization.

Let's focus on a) ensuring that marker (breakpoint) changes do not make the subscribers refresh and b) that the refresh operation initiated by the GitSubscriberMergeContext is not performed on the UI thread.

Patch-set: 2