Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(2 inline comments)

I'm still seeing issues with file renames not being picked up. If I add a file to the file system and then refresh it in Eclipse to get it to show up, it does not show up in the 'Synchronize' view.

Patch-set: 2