[compare] Display revision author instead of HEAD author
The Compare with HEAD action previously showed the SHA-1 and
author of the HEAD commit in the right-side of the compare editor.
The author was therefore associated with the content of the
right-side revision although the HEAD commit may not have actually
modified the file being compared.

This change now displays the SHA-1 and author for the commit where
the file being compared was last modified starting at the commit
that HEAD currently point to.

Change-Id: I035e128cb91568abe3abbc7035237316f1fdd5e9
Signed-off-by: Kevin Sawicki <kevin@github.com>