Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

I like this feature

When having a narrow repositories view (I use to move it to the left side of the workbench right under the package explorer) the text isn't using the available space nicely as the minimum width is a bit too large. Would look nicer if it would reflow the title text and center the links into the middle of the available space until the view really becomes too narrow to show all text.

Patch-set: 4