Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

With this change the file list will display all files which were changed between the selected commit and its parent in the *filtered* view. This is not what the user would expect, and not what the reporter of the bug suggested.

I guess that the suggested solution would be more complicated because you would somehow have to retrieve the real parent additionally to the one from the filtered walk.

Patch-set: 1