Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Works well when there is no rename.

When there is a rename, the TreeFilter is replaced in GitHistoryPage.updateFollowFilter and only the renamed file is shown in the file list.

Maybe this method is not needed anymore at all because now all the files are shown. Then all the great work done in JGits commit 98d4bd6d36d98940f5fd0b6a3e20147cb96903c0 and the FollowingFilter would not be used anymore...

Patch-set: 3