Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

There seems to be an inconsistency here. If I select a branch that's under 'Remote Tracking', the 'Delete' button is enabled. However, when I select both a branch under 'Local' and a branch under 'Remote Tracking', the 'Delete' button becomes disabled.

The user should also not be allowed to delete the branch that is currently checked out.

Patch-set: 3