Update patch set 4
Patch Set 4:

The problem lies in the commit dialog that only sees changes within projects and therefore not removed files in the old location of the projects. The move/delete hook updates the
index as it should, but the commit dialog does not follow
the standard Git process for committing. That has to be
addressed separately.

Patch-set: 4