Update patch set 12
Patch Set 12:

Still RFC-status. We need more tests and fix moving the repository. As I see it the Repository API has to change somehow for this to work, i.e. we must be able to disconnect all users or (option 2) verify that there are no users and abort if there are.

Patch-set: 12