Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

(4 inline comments)

I tried this with a clone of jgit repo and when moving one of the projects one level down, e.g. org.eclipse.jgit to x/org.eclipse.jgit
this seems to work at first glance, but when committing this change only the additions for the new location will be committed, but the removal of the old resources is missing. Looks like these removals have to be staged in the move delete hook.

Patch-set: 4