Update patch set 8
Patch Set 8:

the tests run ok from Eclipse but don't from Maven. I suspect this is caused by the fact that from maven the tests don't run on a full Eclipse platform but on a very reduced platform, so probably we have to enhance the set of bundles used to setup the test environment, see the ui test's pom how this could be done. Another option would be to define an explicit target platform like what we do in egit-github

Patch-set: 8