Do not push all branches from push wizard
The first page of the push wizards allows to choose a remote. The finish
button is enabled to allow for pushing directly to this remote using the
existing refspec. If no refspec exists we should not push all branches
but use the default of JGit's push command, which is (at the moment) to
push the currently checked out branch.

Bug: 363904
Change-Id: Idb3e94d532f13bf7f16a7c07e5f36f166252fb2e
Signed-off-by: Stefan Lay <stefan.lay@sap.com>