Update patch set 6
Patch Set 6: I would prefer that you didn't submit this

That's great if you use the clone wizard alone.

But when using the import wizard "Git -> projects from git" then you can import the projects both in the clone wizard and later on again in the import wizard. I think this is puzzling for the user.

I would suggest to make the clone wizard configurable such that the new project import field is not shown when started from the import wizard.

Patch-set: 6