Always set repository name and branch in decoratable
Project label decorations appear as '[ ]' when the
repository is currently being indexed and the index
diff cache returns a null entry for the repository.

This change sets the repository name and branch before the
index diff data is considered in the DecoratableResourceAdapter
class.

It also updates checks for null with checks for empty 
string in cases where the string can never be null in
the GitLightweightDecorator class.

Bug: 364187
Change-Id: If841e5ac6ffaee38b1734b1d94efa47ec55193d6
Signed-off-by: Kevin Sawicki <kevin@github.com>