Pass file encoding to EditableRevision in FileEditableRevision
Previously, the encoding was not specified. This could lead to spurious
diffs when the default workspace encoding is different from the file
encoding (e.g. for XML files), because the left and right sides were
decoded differently. An example of where this occurred was the Merge
Tool with the HEAD version as left input.

This this change, the problem-causing constructor of EditableRevision
without encoding argument can be deleted, as this was the last code
using it.

Bug: 345160
Change-Id: I88f8bd62a122b3d7c7cc39c5d7242462cdec4038
Signed-off-by: Robin Stocker <robin@nibor.org>