Make ProjectUtils publicly visible
Move the ProjectUtils (and it's required ProjectRecord) to a
non-internal package to enable third-party software to steer on their
own which projects should be imported and connected to GIT.

A real world example would be a plugin that manages it's own list of
"Categories", each containing a set of plugins to import. I have written
a wizard that lets the user chose a category, and it then imports all
associated plugins automatically.

Change-Id: I0dbb5d4a448507c427dedda9f5aad0a00a156528