Add cherry pick action to commit editor toolbar
This moves the bulk of the logic of the cherry pick handler
to a class in the commit.command package and updates the
history handler to forward to the new commit handler.

This also add a confirmation dialog telling the user what
branch the commit will be cherry-picked on top of.

Bug: 360124
Change-Id: Iafb4c1774ab408cb5c3a189ec12c31a94b016670
Signed-off-by: Kevin Sawicki <kevin@github.com>