Update patch set 15
Patch Set 15: I would prefer that you didn't submit this

When opening compare editor on a dirty / not yet staged non-workspace file from staging view in the pane for unstaged changes I can't save changes done in the left pane of the compare editor (corresponding to working tree state of the file). When clicking save or Ctrl-S the change I did in the left pane disappears. When copying a hunk from the right pane to the left using the left arrow button the copying works but the result cannot be saved. For workspace files this works properly.

Patch-set: 15