Make merge and rebase result status more visible
The dialogs displaying the results of a merge or rebase should make the
user aware that something didn't work out, e.g. when a conflict or
an error occurred.

The dialogs now use the standard messagebox icons (error, info, warning)
to communicate the status.

Also made the pull and merge dialogs look a bit more similar.

Change-Id: Id0b608cb2708a870c29727bf8d0c2556ba42d252