Skip clone for import of project from existing repo
Before, we would abort importing when the target directory already
existed. Now we check if the target directory contains a repository that
was cloned from the same URL and if so, skip the clone step. When the
target exists and is not a repo or not for the same URL, we still abort.

Bug: 366207
Change-Id: I591e86fc910c951437881bc1f93fba4c4d466c50
Signed-off-by: Robin Stocker <robin@nibor.org>