Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

When revert hits a merge conflict I still see the progress dialog open and another error dialog opens at the top of the screen (looks pretty ugly) which says that the revert of commit xyz did not complete successfully but it doesn't disclose this is caused by a conflict which I should resolve, stage to declare conflict resolution and commit to seal the deal. This seems to be caused by the fact that we only check for failing paths but not for the merge result.

Patch-set: 2