Update patch set 2
Patch Set 2:

This commit was only aimed at fixing an NPE, it does not change the appearance of the revert failure dialog when failing paths don't exist but unmerged paths do.

I would be happy to improve it to show what unmerged paths exist and instructions for committing them but it should probably be in a different change since it will required a JGit change to the RevertCommand to expose the unmerged paths.

Patch-set: 2