Update patch set 1
Patch Set 1:

This problem is little more complicated then just saying "find bugs to shout up". sharedDocumentListner conected logic is copoed from internal LocalResourceTypedElement class to proper handle non-workspace files; base base class has setter that wasn't overwritten we need to add this code.

True without this setter everything works fine, but I believe that we are missing some corner cases in with this boolean value is set to true and used. In current implementation it is always false so we can remove all if statements that are checking this value and leave only "else" code blocks

Patch-set: 1