Provide proper scheduling rules for operations locking the Git index
Currently the AddToIndexOperation provides the input resources as
scheduling rule. This is wrong because this suggests that two Add To
Index Jobs can run in parallel if the input resources are not
conflicting and belong to the same Git repository.
Now a scheduling rule is used that locks all projects of a given
repository to prevent Jobs running in parallel that modify the index
of the same repository.
The scheduling rules of other EGit operations modifying the Git index
were changed accordingly.

Bug: 366684
Change-Id: I9486e4a06244ab4235d19d6e58144c4a313ba23d
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>