Recover from lock failed exceptions
This begins the work to prompt the user to delete the .lock
file left over by a crashed process that was holding it.

A new LockOperation is added that wraps an IEGitOperation
and prompts the user to delete the lock failure being reported
in a wrapped CoreException.

This commit only currently modified ResetCommand but if this
approach  works then other commands can begin wrapping
core operations in a LockOperation.

JGit-Dependency: I580166d386126bfb54a318a65253070a6e325936
Bug: 360141
Change-Id: Icdd58cc2d7606b23ca1cac550ca287f94f072312