Update patch set 8
Patch Set 8:

What do you think about the following alternative approach: An action is added to Repositories View that allows the user to unlock the Index.
This implies some extra clicks for the user but it avoids the problems that Jobs send a UI and that operations need to be wrapped by LockOperation.
I think the extra clicks are acceptable because hanging index locks do not occur often.

Patch-set: 8