Update patch set 6
Patch Set 6: I would prefer that you didn't submit this

(2 inline comments)

I am not sure if this approach is compliant to Eclipse guide lines.
Is it a good behavior to send dialogs to a user when executing a Job?
What is the result of a job if the operation failed and the user removed the lock afterwards? If the result is failure then the user will get an additional error pop-up from Eclipse. If the result is success, then the result is a bit strange.

Patch-set: 6