Update patch set 8
Patch Set 8:

> What do you think about the following alternative approach: An action is added to Repositories View that allows the user to unlock the Index. This implies some extra clicks for the user but it avoids the problems that Jobs send a UI and that operations need to be wrapped by LockOperation. I think the extra clicks are acceptable because hanging index locks do not occur often.

I am unsure if adding an action to the repository menu that is very advanced and (hopefully) rarely used is the right move.

If if it was moved to an action, it would still require adding a custom dialog to inform them of the situation since the current dialog is the default one for any exception that is thrown and isn't very helpful about what caused the situation and how it can be remedied.  So if a custom dialog is needed to inform them of the situation then it seems better to have them act on it from that dialog.

I have updated the change to no longer do the prompt in a syncExec so it won't be holding any locks now when the user is informed.

Another remaining piece is whether the lock deletion should be scheduled in a job with a scheduling rule that ensures no other jobs could be running that are performing operations that require an index lock.

Patch-set: 8